Initial checkin of the VBus and Venet driver code.
The vbus directory contains the KMDF-based driver code.  The venet
directory contains the NDIS-based venet driver code.  The bin directory 
contains a couple of DOS batch scripts used for installing the drivers 
on a test/development environment.   

The drivers are being developed on a Windows 7 32bit distro, using the 
7.0.0 WDK available from Microsoft.  Ths intention is to create a set of
drivers that will operate on Vista, Server 2003, WIndows 7, and Server
2008.  These drivers will not operate on XP.

At the present time, the drivers install properly and configure/initialize
in a limited scope.  Due to a bug between alacrityvm and the vbus driver,
vbus does not yet 'talk' to the hypervisor properly.  In order to continue
development (and learning), the vbus driver artificially enumerates a
test device with a HardwareId of "VBUS\test_dev_10".  This allowed development
to continue with the venet driver to get installation and the drivers
stacked correctly.  

What this means is that both drivers will install, and configure themselves
within Windows correctly, but you cannot configure the Venet adapter since
the link state is off.  IOW, Windows thinks the NIC is unplugged.  You can
use device manager to locate the drivers and see all the details.

Both drivers include a simple tracing facility intended solely for
debugging.  The drivers will send simple text strings to COM1 of the
Windows VM, which, if correctly configured, will appear on the stdout of
alacrityvm.   Vbus msgs are prefixed with "VBUS" and venet msgs with
"venet".  

Installing and testing the drivers is completely non-trivial.  The LSG wiki
contains a page describing how to configure and setup VM images, however
there is still more information needed on the page to correctly build and
install these drivers.